producer-streaming-api/requirements.txt
fastapi
uvicorn
pandas
psycopg2-binary
kafka-python